# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Helfer API - community-powered bus detection
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LocationPing {
  userId: string;
  lat: number;
  lng: number;
  speed: number;
  accuracy: number;
  heading: number;
  timestamp: number;
  batteryLevel?: number;
}

export interface LocationResponse {
  ok: boolean;
  points: number;
  inCluster: boolean;
  clusterId?: string;
}

export interface BusCluster {
  clusterId: string;
  lat: number;
  lng: number;
  avgSpeed: number;
  confidence: number;
  estimatedPassengers: number;
  heading: number;
  updatedAt: number;
}

export interface ClustersResponse {
  clusters: BusCluster[];
}

export interface ClusterConfirmation {
  userId: string;
  clusterId: string;
  isBus: boolean;
}

export interface ConfirmResponse {
  ok: boolean;
  points: number;
}

export interface UserPoints {
  userId: string;
  points: number;
  tripsToday: number;
  rank: number;
}

export interface AdminStats {
  activeUsers: number;
  activeClusters: number;
  totalLocationPings: number;
  avgConfidence: number;
}

export type GetUserPointsParams = {
  userId: string;
};