# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Helfer API - community-powered bus detection
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/points:
    get:
      operationId: getUserPoints
      tags: [user]
      summary: Get user reward points
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPoints"
        "404":
          description: User not found
...
components:
  schemas:
...
    UserPoints:
      type: object
      properties:
        userId:
          type: string
        points:
          type: number
        tripsToday:
          type: number
        rank:
          type: number
      required:
        - userId
        - points
        - tripsToday
        - rank
...