# lib/api-zod/src/generated/types/userPoints.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Helfer API - community-powered bus detection
 * OpenAPI spec version: 0.1.0
 */

export interface UserPoints {
  userId: string;
  points: number;
  tripsToday: number;
  rank: number;
}